pyproject.toml
[project]
name = "py-discord-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "discord.py>=2.4",
  "python-dotenv>=1.0"
]

[tool.pyright]
typeCheckingMode = "standard"
